NMBD changed

+++ Planner/Domain/Reports/PublicationForm11.cs
@@ -12,7 +12,7 @@ namespace Domain.Reports
 		public String Id { get; set; }
 		public String Name { get; set; }
 		public String FilePath { get; set; }
-		public int Pages { get; set; }
+		public Double Pages { get; set; }
 		public String StoringType { get; set; }
 		public String PublicationType { get; set; }
 		public String Output { get; set; }
